refactor backend handlers into separate modules
