+++ backend/src/handlers/mod.rs
@@ -1,0 +1,7 @@
+pub mod inode;
+pub mod intro;
+pub mod search;
+
+pub use inode::{inode, inode_root};
+pub use intro::{find_name, intro};
+pub use search::{conbine_search, search};
